Add test for QSVT.
