+++ tests/test_execute.py
@@ -87,40 +87,3 @@ class TestQChem:
 
         assert np.allclose(grad_dev_l, grad_qml_l, tol)
         assert np.allclose(grad_dev_l, grad_qml_d, tol)
-
-
-class TestGrover:
-    """Test Grover's algorithm (multi-controlled gates, decomposition, etc.)"""
-
-    @pytest.mark.parametrize("num_qubits", range(4, 8))
-    def test_grover(self, num_qubits):
-        np.random.seed(42)
-        omega = np.random.rand(num_qubits) > 0.5
-        dev = qml.device(device_name, wires=num_qubits)
-        wires = list(range(num_qubits))
-
-        @qml.qnode(dev, diff_method=None)
-        def circuit(omega):
-            iterations = int(np.round(np.sqrt(2**num_qubits) * np.pi / 4))
-
-            # Initial state preparation
-            for wire in wires:
-                qml.Hadamard(wires=wire)
-
-            # Grover's iterator
-            for _ in range(iterations):
-                qml.FlipSign(omega, wires=wires)
-                qml.templates.GroverOperator(wires)
-
-            return qml.probs(wires=wires)
-
-        prob = circuit(omega)
-        index = omega.astype(int)
-        index = functools.reduce(
-            lambda sum, x: sum + (1 << x[0]) * x[1],
-            zip([i for i in range(len(index) - 1, -1, -1)], index),
-            0,
-        )
-        assert np.allclose(np.sum(prob), 1.0)
-        assert prob[index] > 0.95
-        assert np.sum(prob) - prob[index] < 0.05
